Fix spacing issues in suggestions
The suggestions for several issues had bad spacing because of the use of ToInputFormString, which does not try to look good.

Bring in CodeFormat to make the strings look good.

Here are the various issues that are affected:

To test, do CodeInspect with these inputs and see that the suggestions now have proper spacing

"a -> b &"

"a?b&"

"a?b[#]&"

"a_:b_"

"j ~~ a | b ~~ h"