Added the LinterUI.wl kernel file and added a "Kernel" extension to the PacletInfo.wl.in file.
LinterUI.wl:

- This file adds the context `` LinterUI` ``, which contains one non-private function, `ObjectAnalyze`.
- `ObjectAnalyze` runs ``CodeInspector`CodeInspectBox`` on each "Input" and "Code" cell in its argument (a notebook object or a list of cell objects), and attaches a lint pod underneath each one that produced an issue(s).

- The lint pod contains:
	- A marked-up copy of the code, in which hovering over an issue pops a lint raft, containing various actions/info.
	- A column (referred to as a "mooring" in the wl file) of all the lint rafts. Hovering over an in-place lint in the code highlights the corresponding raft in the mooring, and hovering over a mooring raft highlights the in-place issue in the code above.
	- A footer bar, containing information about the number of lint rafts, and an "Apply Edits" button.

- When you take an action on the code, such as clicking on a suggestion to replace an expression, the marked-up code changes to reflect this. Then, you can use the "Apply Edits" button to remove the lint pod and write your changes into the original Input/Code cell.

Also, a small button that indicates a cell has been linted is attached to the top of the cell bracket.

This functionality is intended to be released as a paclet update, just after 12.3 is released. There will be a sister paclet that updates ContextMenus.tr and MenuSetup.tr (and related) files, adding "Analyze Notebook" and "Analyze Cells" menu items.

An inspection palette should accompany the lint pods, that lists out all of the linted cells in a notebook. Version 1 of this paclet will not contain the palette.

This commit does *not* update the cmake files appropriately in the codeinspector repo.