This commit includes some bug fixes, and some groundwork for the Tag Suppression Toggle interface.
## Docked Cell:

Fixed a bug where the analysis summary menu in the docked cell would pop about 30pts beneath the summary button on Windows. This has been fixed by recouching the menu as an ActionMenu rather than an attached cell.

Fixed a sizing issue with the exclam icons in the analysis summary button.

## Lint Pods:

Changed the lint tag Enabled option structure such that tag arguments are now listed in a "Subtag" option in the tag options. For example,
```
<|[tag] -> <|[subtag] -> <|Enabled -> False|>|>|>
```
is now
```
<|[tag] -> <|"Subtags" -> <|[subtag] -> <|Enabled -> False|>|>|>|>
```

The is a subtle behaviour with Association-structured option inheritance in which:
`CurrentValue[scope, opt]` is meant to return the option value set at that scope, and `AbsoluteCurrentValue[scope, opt]` is meant to resolve any `Inherited` values.
However, if an option at the cell level hasn't been set, and is thus `Inherited` by implication but has not been *explicitly* set to the symbol `Inherited`, then CurrentValue behaves as AbsoluteCurrentValue would and resolves the inheritance at the cell level, rather than returning `Inherited`.
This is not the behaviour we want -- we want to be able to see the explicit and inherited option values at any scope, such that we have full control over inheritance in the Tag Suppression Toggle dialog.
Therefore, in `makeRaftMenuIgnoreItem`, before setting tag suppression at the notebook level, there's now a check to see if the option has been set explicitly at the cell level. If it hasn't, then it gets explicitly set to `Inherited` *before* the notebook level option is set.

Note that other cells in the notebook are still going to be contaminated by the notebook-level option, but if you're viewing the $FrontEnd, notebook, and cell option values in the Tag Suppression Toggle dialog, then the cell option values aren't going to be unexpectedly changing when toggling notebook-level options.

-----

Fixed a bug where if CodeInspectBox returned unexpected values, the code pane would pinkbox. Now any such lints are discarded (change to `refineSources`).

In progress: Fixing a bug where the code pane would line-break incorrectly. This involves converting the code pane from a RowBox to an inline Cell (which is guaranteed to display its BoxData as the original input cell would, unlike wrapping the BoxData in a RowBox). However, due to bug 407314, I'll need to do the attached cell deletion manually. See changes in `markupCode`.

## Tag Suppression Dialog

Started adding groundwork for the dialog in the "Tag Suppression Dialog" section of LinterUI.wl.

## Known issues with this commit.

The presence of the docked cell significantly slows down lint pod responsiveness.

Lint pods themselves also slow down the general notebook responsiviness (such as multi-click selection).

Lint rafts don't disappear on mouseout — you have to click outside the lint raft to make it disappear.